refactor(app): move components to each component folder
move components to each component folder
