+++ app/src/organisms/Desktop/ProtocolVisualization/LabwareSlotContainer/index.tsx
@@ -16,9 +16,9 @@ import {
 } from '@opentrons/step-generation'
 
 import styles from './labwareslotcontainer.module.css'
-import { getAllWellContentsAtFrame } from './utils/getAllWellContentsAtFrame'
-import { WellContainer } from './WellContainer'
-import { WellTooltip } from './WellTooltip'
+import { getAllWellContentsAtFrame } from '../utils/getAllWellContentsAtFrame'
+import { WellContainer } from '../WellContainer'
+import { WellTooltip } from '../WellTooltip'
 
 import type { WellGroup } from '@opentrons/components'
 import type { Liquid, RunTimeCommand } from '@opentrons/shared-data'
