+++ app/src/organisms/Desktop/ProtocolVisualization/WellContainer/index.tsx
@@ -9,11 +9,11 @@ import {
   POSITION_REFERENCE_TOP,
 } from '@opentrons/shared-data'
 
-import { TipSvg } from './TipSvg'
-import { getTipSvgInfo } from './utils/getTipSvgInfo'
-import { getWellVolume } from './utils/getWellVolume'
+import { TipSvg } from '../TipSvg'
+import { getTipSvgInfo } from '../utils/getTipSvgInfo'
+import { getWellVolume } from '../utils/getWellVolume'
 import styles from './wellcontainer.module.css'
-import { WellSvg } from './WellSvg'
+import { WellSvg } from '../WellSvg'
 
 import type {
   LabwareWellMap,
